// src/schemas/Temperatures.ts
// Generated by https://github.com/foxglove/foxglove-sdk
// Options: {}

import { Header } from "@/schemas/Header";

/** Temperature readings */
export type Temperatures = {
  /** Standard ROS header */
  header: Header;

  /** Steering motor inverter temp (degC) */
  steering_motor_inverter_temp: number;

  /** Steering motor temp (degC) */
  steering_motor_temp: number;

  /** Curtis controller temp (degC) */
  curtis_controller_temp: number;

  /** Curtis motor temp (degC) */
  curtis_motor_temp: number;
};